requirements.txt
langchain
langchain-community
chromadb
requests
beautifulsoup4
unstructured
python-dotenv
ollama
